Fix a few signed/unsigned comparison warnings.
git-svn-id: svn://svn.minisip.org/minisip/trunk@2236 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libminisip/source/video/codec/AVCoder.cxx
@@ -55,7 +55,7 @@ void rtpCallback( struct AVCodecContext * context, void *data,
 //	memcpy( encoder->rtpPayload + 4, data, size );
 
         encoder->mbCounter += packetNumber;
-        bool endOfFrame = ( encoder->mbCounter == 
+        bool endOfFrame = ( (int32_t)encoder->mbCounter == 
                         ((context->width+15)/16)*((context->height+15)/16) );
 #if 0
         cerr << "NB_MACROBLOCK: " << packetNumber << endl;
